A simple class for storing time-of-day values

The values are stored as a difftime vector with a custom class, and always with "seconds" as unit for robust coercion to numeric. Supports construction from time values, coercion to and from various data types, and formatting. Can be used as a regular column in a data frame.

hms() is a high-level constructor that accepts second, minute, hour and day components as numeric vectors.

new_hms() is a low-level constructor that only checks that its input has the correct base type, numeric.

is_hms() checks if an object is of class hms.

as_hms() forwards to vec_cast().

hms(seconds = NULL, minutes = NULL, hours = NULL, days = NULL)

new_hms(x = numeric())

is_hms(x)

as_hms(x)

# S3 method for hms
as.POSIXct(x, ...)

# S3 method for hms
as.POSIXlt(x, ...)

# S3 method for hms
as.character(x, ...)

# S3 method for hms
as.data.frame(x, row.names = NULL, optional = FALSE, ...,
  nm = paste(deparse(substitute(x), width.cutoff = 500L), collapse =
  " "))

# S3 method for hms
format(x, ...)

# S3 method for hms
print(x, ...)

Arguments

seconds, minutes, hours, days
Time since midnight. No bounds checking is performed.

x
An object.

...
additional arguments to be passed to or from methods.

row.names
NULL or a character vector giving the row names for the data frame. Missing values are not allowed.

optional
logical. If TRUE, setting row names and converting column names (to syntactic names: see make.names) is optional. Note that all of R's base package as.data.frame() methods use optional only for column names treatment, basically with the meaning of data.frame(*, check.names = !optional). See also the make.names argument of the matrix method.

nm
Name of column in new data frame

Details

For hms, all arguments must have the same length or be NULL. Odd combinations (e.g., passing only seconds and hours but not minutes) are rejected.

For arguments of type POSIXct and POSIXlt, as_hms() does not perform timezone conversion. Use lubridate::with_tz() and lubridate::force_tz() as necessary.

Examples

hms(56, 34, 12)

#> 12:34:56

hms()

#> hms()

new_hms(as.numeric(1:3))

#> 00:00:01
#> 00:00:02
#> 00:00:03

# Supports numeric only!
try(new_hms(1:3))

#> Error : `x` must be a vector with type <double>.
#> Instead, it has type <integer>.
#> Backtrace:
#>      █
#>   1. ├─base::tryCatch(...)
#>   2. │ └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>   3. │   ├─base:::tryCatchOne(...)
#>   4. │   │ └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   5. │   └─base:::tryCatchList(expr, names[-nh], parentenv, handlers[-nh])
#>   6. │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>   7. │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>   8. ├─base::withCallingHandlers(...)
#>   9. ├─base::saveRDS(...)
#>  10. ├─base::do.call(...)
#>  11. ├─(function (what, args, quote = FALSE, envir = parent.frame()) ...
#>  12. └─(function (..., crayon_enabled, crayon_colors, pkgdown_internet) ...
#>  13.   └─pkgdown::build_site(...)
#>  14.     └─pkgdown:::build_site_local(...)
#>  15.       └─pkgdown::build_reference(...)
#>  16.         └─purrr::map(...)
#>  17.           └─pkgdown:::.f(.x[[i]], ...)
#>  18.             └─pkgdown:::data_reference_topic(...)
#>  19.               └─pkgdown:::run_examples(...)
#>  20.                 └─pkgdown:::highlight_examples(code, topic, env = env)
#>  21.                   └─evaluate::evaluate(x, child_env(env), new_device = TRUE)
#>  22.                     └─evaluate:::evaluate_call(...)
#>  23.                       ├─evaluate:::timing_fn(...)
#>  24.                       ├─evaluate:::handle(...)
#>  25.                       │ └─base::try(f, silent = TRUE)
#>  26.                       │   └─base::tryCatch(...)
#>  27.                       │     └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  28.                       │       └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  29.                       │         └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  30.                       ├─base::withCallingHandlers(...)
#>  31.                       ├─base::withVisible(eval(expr, envir, enclos))
#>  32.                       └─base::eval(expr, envir, enclos)
#>  33.                         └─base::eval(expr, envir, enclos)
#>  34.                           ├─base::try(new_hms(1:3))
#>  35.                           │ └─base::tryCatch(...)
#>  36.                           │   └─base:::tryCatchList(expr, classes, parentenv, handlers)
#>  37.                           │     └─base:::tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  38.                           │       └─base:::doTryCatch(return(expr), name, parentenv, handler)
#>  39.                           └─hms::new_hms(1:3)
#>  40.                             └─vctrs::vec_assert(x, numeric())

as_hms(1)

#> 00:00:01

as_hms("12:34:56")

#> 12:34:56

as_hms(Sys.time())

#> 20:44:16.739564

as.POSIXct(hms(1))

#> [1] "1970-01-01 00:00:01 UTC"

data.frame(a = hms(1))

#>          a
#> 1 00:00:01

d <- data.frame(hours = 1:3)
d$hours <- hms(hours = d$hours)
d

#>      hours
#> 1 01:00:00
#> 2 02:00:00
#> 3 03:00:00