vec_ptype2.hms {hms}	R Documentation

Coercion

Description

Double dispatch methods to support vctrs::vec_ptype2().

Usage

## S3 method for class 'hms'
vec_ptype2(x, y, ..., x_arg = "", y_arg = "")

Arguments

x
Vector types.

y
Vector types.

...
These dots are for future extensions and must be empty.

x_arg
Argument names for x and y. These are used in error messages to inform the user about the locations of incompatible types (see stop_incompatible_type()).

y_arg
Argument names for x and y. These are used in error messages to inform the user about the locations of incompatible types (see stop_incompatible_type()).

[Package hms version 1.1.1 Index]